Consistent error messages if PK/FK values were invalid
Invalid values are replaced with a Hash with the single key :invalid

If there are PK (duplicate) or FK lookup errors, the returned error
message includes the (stringified) Hash rather than the value that was
in the csv itself, leaking an internal representation which results in
inconsistencies with the way other errors are presented.

Use the value that was invalid rather than the validated result for
consistency.