Add optional dateCheckPattern with datePattern
In field.rb the datePattern field gets used as a strptime format to
parse the date/time. Then it gets used again as a strftime format to
format the DateTime object back out to a string, and a validation error
is thrown if these don't match.

The trouble is, strptime and strftime are not symmetrical.

eg. %H:%M will parse 6:02 and 06:02, but will always format it back out to
06:02

We can use %-H:%M in this case if we want it to format back to 6:02, but
that's not a valid strptime format.

This patch adds an optional dateCheckPattern that is used with strftime.
If it doesn't exist, we fall back to datePattern as before.