Open file in utf-8 + bom mode
This seems to work for files that have a utf8 bom mark, and also for
those that don't. Not sure if this is the right way to do this.

Without this, there's some really hard to work out errors - it says that
"Foo" doesn't match "Foo" as the column name, but what's missing is the
bom bytes don't print out, and the "found" column name includes those.