Method to compute cartesian product of FK values
Its possible to define a column in table that both uses a separator
and is a foreign key to another table. In those cases, we need to check
if every value (after the splitting on the separator) has a matching
value in the other table.

Tables may have compound foreign keys too.

When tables have both compound foreign keys and one of those columns has
a separator, we need to compute the cartesian product of the column
values to later check they exist in the foreign/remote table.

Add a method to compute the cartesian product of column values.