Better way to detect trailing garbage in datetimes
The date/time field parsers have some extra checks in place, which I
believe is to verify that there's no trailing garbage after the
date/time string.

The Ruby strptime will throw an error if there's leading garbage before
the date/time string, but will ignore trailing garbage.

The existing code uses strftime to format the parsed date/time back to a
string, and then compares it to the original. My guess is that is
intended to detect trailing garbage.

The problem is that strptime and strftime are not symmetrical. The
format strings for strftime allow %X for leading zeroes, and %-X for no
leading zeroes. The strptime format allows only %X, and it ignores
leading zeroes.

Previously we added dateCheckFormat so that we could have finer control
over what the check string looked like. This solves the problem, but
we're left with very brittle date parsing, and this causes endless
problems with Excel-exported csv files.

I've discovered that Ruby has a _strptime method alongside strptime.
Instead of returning a date/time object it returns a hash with the
parsed fields. If there is any trailing garbage, that is included in the
has in the `:leftover`. field.

We use this to perform the secondary check - if there is a `:leftover`
field we throw an error.

This lets us be more flexible with the input dates, and still detects
extra leading/trailing garbage, without the extra complexity (and
mystery) of the dateCheckFormat field.